Add mdx, use src folder, remove prisma
