+++ package.json
@@ -15,9 +15,10 @@
     "@fortawesome/free-brands-svg-icons": "^5.15.1",
     "@fortawesome/pro-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.15",
-    "@prisma/client": "^2.29.0",
     "aws-sdk": "^2.778.0",
+    "esbuild": "^0.18.6",
     "framer-motion": "^2.7.7",
+    "mdx-bundler": "^9.2.1",
     "next": "^11.1.0",
     "polished": "^4.1.0",
     "react": "^17.0.2",
@@ -31,8 +32,6 @@
     "@types/styled-components": "^5.1.9",
     "eslint": "^7.23.0",
     "eslint-plugin-react": "^7.23.1",
-    "jest": "^27.0.6",
-    "prisma": "^2.29.0",
     "typescript": "^4.3.5"
   }
 }
