+++ pages/about.tsx
@@ -1,8 +1,8 @@
-import PageTitle from "components/content/PageTitle";
-import Paragraph from "components/content/Paragraph";
-import Section from "components/content/Section";
-import Title from "components/content/Title";
-import Page from "components/layout/Page";
+import PageTitle from "src/components/content/PageTitle";
+import Paragraph from "src/components/content/Paragraph";
+import Section from "src/components/content/Section";
+import Title from "src/components/content/Title";
+import Page from "src/components/layout/Page";
 
 const AboutPage = () => {
   return (
