+++ pages/mdxTest.tsx
@@ -1,0 +1,5 @@
+const mdxTest = () => {
+  return <>sup</>;
+};
+
+export default mdxTest;
