+++ src/components/content/TechStackIcon.tsx
