+++ src/components/global/Springboard/MobileBurger/BurgerLottieJson.tsx
