+++ src/util/hooks/useWindowDimensions.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from "react";
 
-function getWindowDimensions() {
+const getWindowDimensions = () => {
   const { innerWidth: width, innerHeight: height } = window;
   return {
     width,
     height,
   };
-}
+};
 
 interface DimensionState {
   width: number;
